refactor(project): :recycle: rename + clean files
made some better changes for project's names and files organization
