+++ src/main.ts
@@ -2,7 +2,7 @@ import { Notice, Plugin } from "obsidian";
 import {
 	SyncthingController,
 	SyncthingControllerImpl,
-} from "./controllers/syncthing_controller";
+} from "./controllers/main_controller";
 import {
 	SyncThingFromCLI,
 	SyncThingFromCLIimpl,
@@ -11,9 +11,9 @@ import {
 	SyncThingFromREST,
 	SyncThingFromRESTimpl,
 } from "./data/syncthing_remote_datasource";
-import { SyncThingConfiguration } from "./models/syncthing_entities";
-import { ConflictsModal } from "./views/syncthing_conflicts";
-import { SampleSettingTab } from "./views/syncthing_settings_page";
+import { SyncThingConfiguration } from "./models/entities";
+import { ConflictsModal } from "./views/conflicts_modal";
+import { SyncthingSettingTab } from "./views/settings_tab";
 import {
 	DevModeModal,
 	PluginDevModeController,
@@ -21,23 +21,23 @@ import {
 
 //! Remember to rename these classes and interfaces!
 
-interface MyPluginSettings {
+interface SyncthingPluginSettings {
 	api_key: string;
 	configuration: SyncThingConfiguration | Partial<SyncThingConfiguration>;
 	devMode: boolean;
 }
 
-const DEFAULT_SETTINGS: Partial<MyPluginSettings> = {
+const DEFAULT_SETTINGS: Partial<SyncthingPluginSettings> = {
 	configuration: { syncthingBaseUrl: "http://localhost:8384" },
 	devMode: false,
 };
 
-export default class MyPlugin extends Plugin {
+export default class SyncthingPlugin extends Plugin {
 	static loadCount = 0;
-	settings!: MyPluginSettings;
+	settings!: SyncthingPluginSettings;
 	pluginsElements: HTMLElement[] = [];
 	syncthingFromCLI: SyncThingFromCLI = new SyncThingFromCLIimpl();
-	syncthingFromREST: SyncThingFromREST = new SyncThingFromRESTimpl();
+	syncthingFromREST: SyncThingFromREST = new SyncThingFromRESTimpl(this);
 	syncthingController: SyncthingController = new SyncthingControllerImpl(
 		this.syncthingFromCLI,
 		this.syncthingFromREST,
@@ -48,7 +48,7 @@ export default class MyPlugin extends Plugin {
 	);
 
 	async onload() {
-		MyPlugin.loadCount++;
+		SyncthingPlugin.loadCount++;
 		await this.loadSettings();
 
 		// This adds a status bar item to the bottom of the app. Does not work on mobile apps.
@@ -58,9 +58,10 @@ export default class MyPlugin extends Plugin {
 			new Notice("SyncThing integration is not yet implemented.");
 		});
 
-		const pluginSettingTab = new SampleSettingTab(this.app, this);
+		const pluginSettingTab = new SyncthingSettingTab(this.app, this);
 		// This adds a settings tab so the user can configure various aspects of the plugin
-		if (MyPlugin.loadCount === 1) this.addSettingTab(pluginSettingTab);
+		if (SyncthingPlugin.loadCount === 1)
+			this.addSettingTab(pluginSettingTab);
 
 		const syncthingConflictManager = this.addRibbonIcon(
 			"construction",
